Major organizational changes, improved code, and more
