+++ AscentsPlusMore/AscentsTables.py
@@ -1,57 +1,16 @@
-from AllPermutationsLexicographic import allperms
-from NumAscents import numAscents
-from tValueCalculator import tval
+from MiscellaneousTools.FindIndices import findIndices
+from NumAscents import *
+from StartsEndsWith import *
+from tValueCalculator import *
 import pandas as pd
-
-
-def numAscentsLex(n):
-    perms = allperms(n)
-    list = []
-    for perm in perms:
-        list.append(numAscents(perm))
-    return list
-
-
-def startsWith(p):
-    return p[0]
-
-
-def startsWithLex(n):
-    perms = allperms(n)
-    list = []
-    for perm in perms:
-        list.append(perm[0])
-    return list
-
-
-def tvals(n):
-    perms = allperms(n)
-    list = []
-    for perm in perms:
-        list.append(tval(perm))
-    return list
-
-
-def tIndices(list, item):
-    x = -1
-    indices = []
-    while True:
-        try:
-            index = list.index(item, x + 1)
-        except ValueError:
-            break
-        else:
-            indices.append(index)
-            x = index
-    return indices
 
 
 def ascentSSTable(n):
     ascents = numAscentsLex(n)
-    tvalues = tvals(n)
+    tvalues = tValues(n)
     columns = []
     for i in range(n):
-        indices = tIndices(ascents, i)
+        indices = findIndices(ascents, i)
         newtvalues = []
         for index in indices:
             newtvalues.append(tvalues[index])
@@ -70,10 +29,10 @@ def ascentSSTable(n):
 
 def startsWithSSTable(n):
     sw = startsWithLex(n)
-    tvalues = tvals(n)
+    tvalues = tValues(n)
     columns = []
     for i in range(1, n + 1):
-        indices = tIndices(sw, i)
+        indices = findIndices(sw, i)
         newtvalues = []
         for index in indices:
             newtvalues.append(tvalues[index])
@@ -91,24 +50,12 @@ def startsWithSSTable(n):
     return table.rename(columns={i: labels[i] for i in range(n)})
 
 
-def endsWith(p):
-    return p[-1]
-
-
-def endsWithLex(n):
-    perms = allperms(n)
-    list = []
-    for perm in perms:
-        list.append(perm[-1])
-    return list
-
-
 def endsWithSSTable(n):
     ew = endsWithLex(n)
-    tvalues = tvals(n)
+    tvalues = tValues(n)
     columns = []
     for i in range(1, n + 1):
-        indices = tIndices(ew, i)
+        indices = findIndices(ew, i)
         newtvalues = []
         for index in indices:
             newtvalues.append(tvalues[index])
