+++ README.md
@@ -1,3 +1,41 @@
 # StackSortingPermutations
 
-GO TO main.py !!!!!!!!
+<h2><u><i> File Directory </i></u></h2>
+
+<ul>
+    <li><b> Dilate: </b> One function dilateP that dilates vertices of a polytope by a given factor</li>
+    <li><b> FundamentalParallelepipedEnumerator: </b> Tools to enumerate all lattice points in the 
+        fundamental parallelepiped of a given polytope (only Ln1 families for now)</li>
+    <li><b> Ln1Permutations: </b> Various functions to generate Ln1 permutations and their families</li>
+    <li><b> SimplexPointEnumerator: </b> Tools to enumerate all lattice points in the a given simplex</li>
+    <li><b> StackSort: </b> The stack-sorting algorithm plus extra functions to quickly stack sort large families of 
+        permutations</li>
+    <li><b> tValueCalculator: </b> Calculate t-values for given permutations, the number of times it takes the stack
+        sorting algorithm to reach the identity</li>
+</ul>
+
+<h3> AscentsPlusMore </h3>
+<ul>
+    <li><b> AscentsTables: </b> Code for the tables in TablesTool, see below</li>
+    <li><b> NumAscents: </b> Calculate the number of ascents in given permutations</li>
+    <li><b> StartsEndsWith: </b> Functions to find the numbers permutations start and end with</li>
+    <li><b> TablesTool: </b> Generates the tables from our prior discussions, compares number of iterations of the 
+        algorithm to ascents, starting number, and ending number for large families of permutations 
+        (used to find patterns)</li>
+</ul>
+
+<h3> MiscellaneousTools </h3>
+<ul>
+    <li><b> FindIndices: </b> One function findIndices, takes a list and an item, outputs a list containing all indices
+        where the item is found in the list</li>
+</ul>
+
+<h3> PermutationTools </h3>
+<ul>
+    <li><b> AllPermutationsLexicographic: </b> Code to enumerate all permutations from 1 to n (given by user) in 
+        lexicographic order</li>
+    <li><b> IdentityPermutation: </b> One function identityp, takes a positive integer n and outputs the identity
+        permutation 1,2,...,n</li>
+</ul>
+
+
